web_ide/requirements.txt
Django==1.8.3
djangorestframework==3.1.3
funcsigs==0.4
gitdb==0.6.4
GitPython==1.0.1
Markdown==2.6.2
mock==1.2.0
pbr==1.3.0
psycopg2==2.6.1
python-memcached==1.54
requests==2.7.0
six==1.9.0
smmap==0.9.0
wheel==0.24.0
